% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_arrow.R
\name{create_metadata}
\alias{create_metadata}
\title{Create standardised geo metadata for Parquet files}
\usage{
create_metadata(df)
}
\arguments{
\item{df}{object of class \code{sf}}
}
\value{
JSON formatted list with geo-metadata
}
\description{
Create standardised geo metadata for Parquet files
}
\details{
Reference for metadata standard:
\url{https://github.com/geopandas/geo-arrow-spec}. This is compatible with
\code{GeoPandas} Parquet files.
}
\keyword{internal}
