% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_arrow.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Basic checking of key geo metadata columns}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{list for geo metadata}
}
\value{
None. Throws an error and stops execution
}
\description{
Basic checking of key geo metadata columns
}
\keyword{internal}
